backend/requirements.txt
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
autopep8==1.6.0
pymysql==1.0.2
mysqlclient==2.1.0
django-cors-headers==3.11.0
factory-boy==3.2.1
faker_vehicle
cryptography